requirements.txt
botbuilder-integration-aiohttp>=4.14.0
langchain
openai
langchain-openai
requests
langchain-chroma